+++ x/data/server/query_attestations_by_attestor.go
@@ -7,7 +7,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	api "github.com/regen-network/regen-ledger/api/regen/data/v1"
-	errors "github.com/regen-network/regen-ledger/errors"
+	regenerrors "github.com/regen-network/regen-ledger/errors"
 	"github.com/regen-network/regen-ledger/types"
 	"github.com/regen-network/regen-ledger/types/ormutil"
 	"github.com/regen-network/regen-ledger/x/data"
@@ -16,17 +16,17 @@ import (
 // AttestationsByAttestor queries data attestations by an attestor.
 func (s serverImpl) AttestationsByAttestor(ctx context.Context, request *data.QueryAttestationsByAttestorRequest) (*data.QueryAttestationsByAttestorResponse, error) {
 	if len(request.Attestor) == 0 {
-		return nil, errors.ErrInvalidArgument.Wrap("attestor cannot be empty")
+		return nil, regenerrors.ErrInvalidArgument.Wrap("attestor cannot be empty")
 	}
 
 	addr, err := sdk.AccAddressFromBech32(request.Attestor)
 	if err != nil {
-		return nil, errors.ErrInvalidArgument.Wrapf("attestor: %s", err.Error())
+		return nil, regenerrors.ErrInvalidArgument.Wrapf("attestor: %s", err.Error())
 	}
 
 	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
 	if err != nil {
-		return nil, errors.ErrInvalidArgument.Wrap(err.Error())
+		return nil, regenerrors.ErrInvalidArgument.Wrap(err.Error())
 	}
 
 	it, err := s.stateStore.DataAttestorTable().List(
@@ -48,7 +48,7 @@ func (s serverImpl) AttestationsByAttestor(ctx context.Context, request *data.Qu
 
 		dataID, err := s.stateStore.DataIDTable().Get(ctx, dataAttestor.Id)
 		if err != nil {
-			return nil, errors.ErrNotFound.Wrap(err.Error())
+			return nil, regenerrors.ErrNotFound.Wrap(err.Error())
 		}
 
 		attestations = append(attestations, &data.AttestationInfo{
@@ -60,7 +60,7 @@ func (s serverImpl) AttestationsByAttestor(ctx context.Context, request *data.Qu
 
 	pageRes, err := ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
 	if err != nil {
-		return nil, errors.ErrInternal.Wrap(err.Error())
+		return nil, regenerrors.ErrInternal.Wrap(err.Error())
 	}
 
 	return &data.QueryAttestationsByAttestorResponse{