+++ x/data/server/query_attestations_by_iri.go
@@ -7,7 +7,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	api "github.com/regen-network/regen-ledger/api/regen/data/v1"
-	errors "github.com/regen-network/regen-ledger/errors"
+	regenerrors "github.com/regen-network/regen-ledger/errors"
 	"github.com/regen-network/regen-ledger/types"
 	"github.com/regen-network/regen-ledger/types/ormutil"
 	"github.com/regen-network/regen-ledger/x/data"
@@ -16,23 +16,23 @@ import (
 // AttestationsByIRI queries data attestations by the IRI of the data.
 func (s serverImpl) AttestationsByIRI(ctx context.Context, request *data.QueryAttestationsByIRIRequest) (*data.QueryAttestationsByIRIResponse, error) {
 	if len(request.Iri) == 0 {
-		return nil, errors.ErrInvalidArgument.Wrap("IRI cannot be empty")
+		return nil, regenerrors.ErrInvalidArgument.Wrap("IRI cannot be empty")
 	}
 
 	// check for valid IRI
 	_, err := data.ParseIRI(request.Iri)
 	if err != nil {
-		return nil, errors.ErrInvalidArgument.Wrapf("failed to parse IRI: %s", err.Error())
+		return nil, regenerrors.ErrInvalidArgument.Wrapf("failed to parse IRI: %s", err.Error())
 	}
 
 	dataID, err := s.stateStore.DataIDTable().GetByIri(ctx, request.Iri)
 	if err != nil {
-		return nil, errors.ErrNotFound.Wrapf("data record with IRI: %s", request.Iri)
+		return nil, regenerrors.ErrNotFound.Wrapf("data record with IRI: %s", request.Iri)
 	}
 
 	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
 	if err != nil {
-		return nil, errors.ErrInvalidArgument.Wrap(err.Error())
+		return nil, regenerrors.ErrInvalidArgument.Wrap(err.Error())
 	}
 
 	it, err := s.stateStore.DataAttestorTable().List(
@@ -61,7 +61,7 @@ func (s serverImpl) AttestationsByIRI(ctx context.Context, request *data.QueryAt
 
 	pageRes, err := ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
 	if err != nil {
-		return nil, errors.ErrInternal.Wrap(err.Error())
+		return nil, regenerrors.ErrInternal.Wrap(err.Error())
 	}
 
 	return &data.QueryAttestationsByIRIResponse{