+++ x/data/server/query_resolver.go
@@ -5,19 +5,19 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	errors "github.com/regen-network/regen-ledger/errors"
+	regenerrors "github.com/regen-network/regen-ledger/errors"
 	"github.com/regen-network/regen-ledger/x/data"
 )
 
 // Resolver queries a resolver by its unique identifier.
 func (s serverImpl) Resolver(ctx context.Context, request *data.QueryResolverRequest) (*data.QueryResolverResponse, error) {
 	if request.Id == 0 {
-		return nil, errors.ErrInvalidArgument.Wrap("ID cannot be empty")
+		return nil, regenerrors.ErrInvalidArgument.Wrap("ID cannot be empty")
 	}
 
 	resolver, err := s.stateStore.ResolverTable().Get(ctx, request.Id)
 	if err != nil {
-		return nil, errors.ErrNotFound.Wrapf("resolver with ID: %d", request.Id)
+		return nil, regenerrors.ErrNotFound.Wrapf("resolver with ID: %d", request.Id)
 	}
 
 	manager := sdk.AccAddress(resolver.Manager).String()